chore(tests): move js tracer into a single module
