+++ tests/it/main.rs
@@ -2,6 +2,8 @@
 pub mod utils;
 
 mod geth;
+#[cfg(feature = "js-tracer")]
+mod geth_js;
 mod parity;
 mod transfer;
 mod writer;
